@@ -18,7 +18,7 @@ authors = ["@shm11C3"]
 license = ""
 repository = ""
 edition = "2024"
-rust-version = "1.85"
+rust-version = "1.88"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
